+++ lbz/exceptions.py
@@ -1,6 +1,7 @@
 # coding=utf-8
 
 from http import HTTPStatus
+from typing import Optional
 
 from lbz.response import Response
 
@@ -14,6 +15,7 @@ class LambdaFWException(Exception):
 
     message = HTTPStatus.INTERNAL_SERVER_ERROR.description
     status_code = HTTPStatus.INTERNAL_SERVER_ERROR.value
+    error_code: Optional[str] = None
 
     def __init__(self, message: str = "") -> None:
         super().__init__(message)
@@ -21,14 +23,17 @@ class LambdaFWException(Exception):
             self.message = message
 
     def __str__(self) -> str:
+        if self.error_code is not None:
+            return f"[{self.status_code}] {self.error_code} - {self.message}"
         return f"[{self.status_code}] {self.message}"
 
     def get_response(self, request_id: str) -> Response:
         """Creates a proper standardised Response for Errors."""
-        return Response(
-            {"message": self.message, "request_id": request_id},
-            status_code=self.status_code,
-        )
+        resp_data = {"message": self.message, "request_id": request_id}
+        if self.error_code:
+            resp_data["error_code"] = self.error_code
+
+        return Response(resp_data, status_code=self.status_code)
 
 
 class LambdaFWClientException(LambdaFWException):
@@ -87,7 +92,7 @@ class UnsupportedMethod(LambdaFWClientException):
     status_code = HTTPStatus.METHOD_NOT_ALLOWED.value
 
     def __init__(self, method: str) -> None:
-        super().__init__(message="Unsupported method: %s" % method)
+        super().__init__(message=f"Unsupported method: {method}")
 
 
 class NotAcceptable(LambdaFWClientException):