+++ tests/test_exceptions.py
@@ -45,14 +45,32 @@ from lbz.exceptions import (
     URITooLong,
     VariantAlsoNegotiates,
 )
-from lbz.response import Response
 
 
 def test_lambda_fw_exception() -> None:
-    exp = LambdaFWException(message="Nope")
-    assert exp.message == "Nope"
-    assert exp.status_code == 500
-    assert isinstance(exp.get_response(request_id=""), Response)
+    exception = LambdaFWException(message="Nope")
+    assert str(exception) == "[500] Nope"
+    assert exception.message == "Nope"
+    assert exception.status_code == 500
+
+    response = exception.get_response(request_id="test-req-id")
+    assert response.body == {"message": "Nope", "request_id": "test-req-id"}
+
+
+def test_lambda_fw_exception_with_error_code() -> None:
+    class TestException(LambdaFWException):
+        error_code = "TEST_ERR"
+
+    exception = TestException(message="Nope")
+    assert str(exception) == "[500] TEST_ERR - Nope"
+    assert exception.message == "Nope"
+    assert exception.error_code == "TEST_ERR"
+    assert exception.status_code == 500
+    assert exception.get_response(request_id="test-req-id").body == {
+        "error_code": "TEST_ERR",
+        "message": "Nope",
+        "request_id": "test-req-id",
+    }
 
 
 def test_unsupported_method() -> None: